+++ src/Http/Requests/ResourceQuery.php
@@ -53,6 +53,13 @@ class ResourceQuery extends FormRequest implements QueryParameters
     ];
 
     /**
+     * The include paths to use if the client provides none.
+     *
+     * @var string[]|null
+     */
+    protected ?array $defaultIncludePaths = null;
+
+    /**
      * Specify the callback to use to guess the request class for querying many resources.
      *
      * @param callable $resolver
@@ -149,7 +156,7 @@ class ResourceQuery extends FormRequest implements QueryParameters
             return IncludePaths::fromString($data['include'] ?: '');
         }
 
-        return null;
+        return IncludePaths::nullable($this->defaultIncludePaths());
     }
 
     /**
@@ -220,6 +227,16 @@ class ResourceQuery extends FormRequest implements QueryParameters
             'page',
             'filter',
         ])->all();
+    }
+
+    /**
+     * Get the default include paths to use if the client has provided none.
+     *
+     * @return string[]|null
+     */
+    protected function defaultIncludePaths(): ?array
+    {
+        return $this->defaultIncludePaths;
     }
 
     /**